# app/Http/Controllers/Api/CourseController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Http\Requests\StoreCourseRequest;
use App\Http\Requests\UpdateCourseRequest;
use App\Models\Course;
use App\Services\CourseService;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use OpenApi\Attributes as OA;

#[OA\Tag(name: 'Course', description: 'Gestion des courses')]
class CourseController extends Controller
{
    public function __construct(
        private readonly CourseService $service
    ) {}

    #[OA\Get(
        path: '/v1/courses',
        summary: 'Liste des courses',
        description: 'Retourne une liste paginée avec filtres optionnels',
        operationId: 'getCourses',
        tags: ['Course'],
        security: [['sanctum' => []]]
    )]
    #[OA\Parameter(name: 'page', in: 'query', description: 'Numéro de page', required: false, schema: new OA\Schema(type: 'integer', default: 1))]
    #[OA\Parameter(name: 'per_page', in: 'query', description: 'Nombre d’éléments par page', required: false, schema: new OA\Schema(type: 'integer', default: 15))]
    #[OA\Parameter(name: 'search', in: 'query', description: 'Recherche par nom', required: false, schema: new OA\Schema(type: 'string'))]
    #[OA\Response(
        response: 200,
        description: 'Liste récupérée avec succès',
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'success', type: 'boolean', example: true),
                new OA\Property(property: 'message', type: 'string'),
                new OA\Property(property: 'data', type: 'array', items: new OA\Items(ref: '#/components/schemas/CourseResource')),
                new OA\Property(property: 'meta', type: 'object'),
            ]
        )
    )]
    #[OA\Response(response: 401, description: 'Non authentifié', content: new OA\JsonContent(ref: '#/components/schemas/ErrorResponse'))]
    public function index(Request $request): JsonResponse
    {
        return $this->service->index($request);
    }

    #[OA\Get(
        path: '/v1/courses/{course}',
        summary: 'Détails d’un(e) Course',
        operationId: 'getCourse',
        tags: ['Course'],
        security: [['sanctum' => []]]
    )]
    #[OA\Parameter(name: 'course', in: 'path', required: true, schema: new OA\Schema(type: 'string', format: 'uuid'))]
    #[OA\Response(
        response: 200,
        description: 'Course trouvé(e)',
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'success', type: 'boolean', example: true),
                new OA\Property(property: 'message', type: 'string'),
                new OA\Property(property: 'data', ref: '#/components/schemas/CourseResource'),
            ]
        )
    )]
    #[OA\Response(response: 404, description: 'Non trouvé(e)', content: new OA\JsonContent(ref: '#/components/schemas/ErrorResponse'))]
    public function show(Course $course): JsonResponse
    {
        return $this->service->show($course);
    }

    #[OA\Post(
        path: '/v1/courses',
        summary: 'Créer un(e) Course',
        operationId: 'createCourse',
        tags: ['Course'],
        security: [['sanctum' => []]]
    )]
    #[OA\RequestBody(required: true, content: new OA\JsonContent(ref: '#/components/schemas/StoreCourseRequest'))]
    #[OA\Response(
        response: 201,
        description: 'Course créé(e)',
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'success', type: 'boolean', example: true),
                new OA\Property(property: 'message', type: 'string'),
                new OA\Property(property: 'data', ref: '#/components/schemas/CourseResource'),
            ]
        )
    )]
    #[OA\Response(response: 422, description: 'Erreur de validation', content: new OA\JsonContent(ref: '#/components/schemas/ValidationErrorResponse'))]
    public function store(StoreCourseRequest $request): JsonResponse
    {
        // $this->authorize('create', Course::class);

        return $this->service->store($request);
    }

    #[OA\Put(
        path: '/v1/courses/{course}',
        summary: 'Mettre à jour un(e) Course',
        operationId: 'updateCourse',
        tags: ['Course'],
        security: [['sanctum' => []]]
    )]
    #[OA\Parameter(name: 'course', in: 'path', required: true, schema: new OA\Schema(type: 'string', format: 'uuid'))]
    #[OA\RequestBody(required: true, content: new OA\JsonContent(ref: '#/components/schemas/UpdateCourseRequest'))]
    #[OA\Response(
        response: 200,
        description: 'Course mis(e) à jour',
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'success', type: 'boolean', example: true),
                new OA\Property(property: 'message', type: 'string'),
                new OA\Property(property: 'data', ref: '#/components/schemas/CourseResource'),
            ]
        )
    )]
    #[OA\Response(response: 404, description: 'Non trouvé(e)', content: new OA\JsonContent(ref: '#/components/schemas/ErrorResponse'))]
    public function update(UpdateCourseRequest $request, Course $course): JsonResponse
    {
        // $this->authorize('update', $course);

        return $this->service->update($course, $request);
    }

    #[OA\Delete(
        path: '/v1/courses/{course}',
        summary: 'Supprimer un(e) Course',
        operationId: 'deleteCourse',
        tags: ['Course'],
        security: [['sanctum' => []]]
    )]
    #[OA\Parameter(name: 'course', in: 'path', required: true, schema: new OA\Schema(type: 'string', format: 'uuid'))]
    #[OA\Response(
        response: 200,
        description: 'Course supprimé(e)',
        content: new OA\JsonContent(
            properties: [
                new OA\Property(property: 'success', type: 'boolean', example: true),
                new OA\Property(property: 'message', type: 'string'),
            ]
        )
    )]
    #[OA\Response(response: 404, description: 'Non trouvé(e)', content: new OA\JsonContent(ref: '#/components/schemas/ErrorResponse'))]
    public function destroy(Course $course): JsonResponse
    {
        // $this->authorize('delete', $course);

        return $this->service->destroy($course);
    }
}