# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OrientTogo API",
        "description": "Documentation officielle de l'API OrientTogo (Laravel 13). \n    Toutes les requêtes doivent être préfixées par `/api/v1`. \n    L'authentification se fait via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Équipe Technique OrientTogo",
            "email": "contact@orienttogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "/api",
            "description": "Serveur Local (Base Path)"
        }
    ],
    "paths": {
...
        "/v1/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Liste des courses",
                "description": "Retourne une liste paginée avec filtres optionnels",
                "operationId": "getCourses",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d’éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Créer un(e) Course",
                "operationId": "createCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course créé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{course}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Détails d’un(e) Course",
                "operationId": "getCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "Mettre à jour un(e) Course",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course mis(e) à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "Supprimer un(e) Course",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course supprimé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ est obligatoire."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StoreCourseRequest": {
                "required": [
                    "establishment_id",
                    "name",
                    "level",
                    "description",
                    "duration_months",
                    "annual_fees",
                    "accreditation"
                ],
                "properties": {
                    "establishment_id": {
                        "type": "string",
                        "example": "Exemple de establishment_id"
                    },
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "level": {
                        "type": "string",
                        "example": "Exemple de level"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "duration_months": {
                        "type": "integer",
                        "example": 42
                    },
                    "annual_fees": {
                        "type": "number",
                        "example": 99.99
                    },
                    "accreditation": {
                        "type": "string",
                        "example": "Exemple de accreditation"
                    }
                },
                "type": "object"
            },
...
            "UpdateCourseRequest": {
                "properties": {
                    "establishment_id": {
                        "type": "string",
                        "example": "Exemple de establishment_id"
                    },
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "level": {
                        "type": "string",
                        "example": "Exemple de level"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "duration_months": {
                        "type": "integer",
                        "example": 42
                    },
                    "annual_fees": {
                        "type": "number",
                        "example": 99.99
                    },
                    "accreditation": {
                        "type": "string",
                        "example": "Exemple de accreditation"
                    }
                },
                "type": "object"
            },
...
            "CourseResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "establishment_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Licence Informatique"
                    },
                    "level": {
                        "type": "string",
                        "example": "licence"
                    },
                    "description": {
                        "type": "string",
                        "example": "..."
                    },
                    "duration_months": {
                        "type": "integer",
                        "example": 36
                    },
                    "annual_fees": {
                        "type": "number",
                        "example": 50000
                    },
                    "accreditation": {
                        "type": "string",
                        "example": "AMESRES"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utilisez le token retourné lors de la connexion (login ou verify OTP).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}